requirements/django-venv-requirements_2-23-2025.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
codegrade==16.1.84
constantly==23.10.4
coverage==7.6.12
cryptography==44.0.0
daphne==4.1.2
dj-rest-auth==7.0.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docformatter==1.7.5
drf-nested-routers==0.94.1
drf-spectacular==0.26.2
flake8==7.1.2
flake8-docstrings==1.7.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
msgpack==1.1.0
mypy-extensions==1.0.0
nodejs==0.1.1
npm==0.1.1
numpy==2.2.3
oauthlib==3.2.2
optional-django==0.1.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
perl==1.0.0
platformdirs==4.3.7
pluggy==1.5.0
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
rsa==4.9
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sqlparse==0.5.3
structlog==20.2.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
untokenize==0.1.1
uritemplate==4.1.1
urllib3==2.3.0
validate_email==1.3
zope.interface==7.2
